Split the Bible catalogue browser by concern
The dialog was already twenty top-level composables with almost no shared
state, so each moved as it stood.

  BibleCatalogBrowserDialog 307, Controls 282, ModuleRow 243,
  Licence 210, Labels 67
